% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{read_config}
\alias{read_config}
\title{Read the scrape configuration}
\usage{
read_config(config = system.file("config", "packages.yml", package = "natmcp"))
}
\arguments{
\item{config}{Path to the scrape config YAML.}
}
\value{
The parsed config as a list.
}
\description{
Read the scrape configuration
}
